Обновление кода из проекта ringoman.
[banners]
 - Обновлён вызов объектов на шаблон разработки "фабрика объектов"
 - Добавлен метод выполняющий выбор баннеров по расчёту коэффициента
 - Обновлён виджет выбора нескольких баннеров

[catsubcat]
 - Исправлены различные ошибки уровня notice
 - Восстановлен некоторый потерянный функционал
 - Убрана поддержка работы в облегчённом режиме

[hints]
 - Добавлен новый модуль hints обеспечивающий вывод редактируемых
   всплывающих подсказок на сайте.

[interfaces]
 - Добавлена возможность фильтрации по диапазонам больше/меньше
 - Добавлен собственный метод контроля пустоты значения вместо empty
 - Исправлены мелкие ошибки уровня notice
 - Удалён лишний код в классе CHTMLParser, добавлен массив со списком допустимых
   тэгов
 - Исправлен метод поиска страницы записи

[main]
 - Добавлена возможность скрыть редактор tinymce в административном интерфейсе
 - Исправлены тексты ошибок
 - Добавлен метод GetAdditionalText для класса CError
 - Обновлён и исправлен код менеджера событий
 - Множество исправлений уровня notice
 - Обновления классов работы с файлами
 - Добавлена операция поиска по началу строки и концу строки в методе GetList
 - Добавлен метод проверки возможности вызова виджета через ajax
 - Добавлен виджет позволяющий вывести значение настройки

[navigation]
 - Исправлена работа с меню в административной части

[redirect]
 - Добавлено управление автоматическим переходом в ПЧ

[sitemap]
 - Исправления мелких ошибок и обновление до 2.6.

[wave]
 - Добавлены тексты уведомлений

+++ modules/main/libs/class.CFilesObject.php
@@ -147,9 +147,13 @@ class CFilesObject extends CObject
 					$arResult[$field]=$obUploadManager->Upload($this->sUploadPath.'/'.$this->_GenFileName($obUploadManager->GetFileName()),false);
 				}
 			}
-			else
+			elseif(isset($data[$prefix.$field]))
 			{
 				$arResult[$field]=$data[$prefix.$field];
+			}
+			else
+			{
+				$arResult[$field]='';
 			}
 		}
 		return $arResult;
@@ -161,6 +165,11 @@ class CFilesObject extends CObject
 	protected function _GenFileName($filename)
 	{
 		return md5($filename.time()).'.'.substr($filename,strrpos($filename,'.')+1);
+	}
+
+	function GenFileName($filename)
+	{
+		return $this->_GenFileName($filename);
 	}
 
 	/**
@@ -189,7 +198,7 @@ class CFilesObject extends CObject
 					if($input[$prefix.'id']!='')
 					{
 						$arItem=$this->GetRecord(array('id'=>$input[$prefix . 'id']));
-						if(is_array($arItem)&&($arItem['id']==$input[$prefix . 'id']))
+						if(is_array($arItem)&&($arItem['id']==$input[$prefix . 'id']) && $sResult!=$arItem[$key])
 						{
 							if (file_exists(UPLOADS_DIR.$arItem[$key])&&is_file(UPLOADS_DIR.$arItem[$key]))
 								unlink(UPLOADS_DIR.$arItem[$key]);
